+++ build_cli/test/src/peanut_example.g.dart
@@ -56,8 +56,7 @@ final _$parserForPeanutOptions = new ArgParser()
       allowed: ['to-source', 'from-source', 'via-assets'])
   ..addFlag('help',
       abbr: 'h', help: 'Prints usage information.', negatable: false)
-  ..addFlag('secret', hide: true)
-  ..addOption('cool-bean');
+  ..addFlag('secret', hide: true);
 PeanutOptions parsePeanutOptions(List<String> args) {
   var result = _$parserForPeanutOptions.parse(args);
   return _$parsePeanutOptionsResult(result);